+++ src/Iterator.php
@@ -10,10 +10,7 @@
 namespace SebastianBergmann\FileIterator;
 
 use function array_any;
-use function preg_match;
-use function realpath;
 use function str_ends_with;
-use function str_replace;
 use function str_starts_with;
 use FilterIterator;
 use SplFileInfo;
@@ -25,10 +22,6 @@ use SplFileInfo;
  */
 final class Iterator extends FilterIterator
 {
-    public const int PREFIX = 0;
-    public const int SUFFIX = 1;
-    private false|string $basePath;
-
     /**
      * @var list<string>
      */
@@ -44,9 +37,8 @@ final class Iterator extends FilterIterator
      * @param list<string>                $suffixes
      * @param list<string>                $prefixes
      */
-    public function __construct(string $basePath, \Iterator $iterator, array $suffixes = [], array $prefixes = [])
+    public function __construct(\Iterator $iterator, array $suffixes = [], array $prefixes = [])
     {
-        $this->basePath = realpath($basePath);
         $this->prefixes = $prefixes;
         $this->suffixes = $suffixes;
 
@@ -55,52 +47,18 @@ final class Iterator extends FilterIterator
 
     public function accept(): bool
     {
-        $current = $this->getInnerIterator()->current();
+        $filename = $this->getInnerIterator()->current()->getFilename();
 
-        $filename = $current->getFilename();
-        $realPath = $current->getRealPath();
-
-        if ($realPath === false) {
-            // @codeCoverageIgnoreStart
+        if ($this->prefixes !== [] &&
+            !array_any($this->prefixes, static fn (string $prefix) => str_starts_with($filename, $prefix))) {
             return false;
-            // @codeCoverageIgnoreEnd
         }
 
-        return $this->acceptPath($realPath) &&
-               $this->acceptPrefix($filename) &&
-               $this->acceptSuffix($filename);
-    }
-
-    private function acceptPath(string $path): bool
-    {
-        // Filter files in hidden directories by checking path that is relative to the base path.
-        if (preg_match('=/\.[^/]*/=', str_replace((string) $this->basePath, '', $path)) === 1) {
+        if ($this->suffixes !== [] &&
+            !array_any($this->suffixes, static fn (string $suffix) => str_ends_with($filename, $suffix))) {
             return false;
         }
 
         return true;
-    }
-
-    private function acceptPrefix(string $filename): bool
-    {
-        return $this->acceptSubString($filename, $this->prefixes, self::PREFIX);
-    }
-
-    private function acceptSuffix(string $filename): bool
-    {
-        return $this->acceptSubString($filename, $this->suffixes, self::SUFFIX);
-    }
-
-    /**
-     * @param list<string> $subStrings
-     */
-    private function acceptSubString(string $filename, array $subStrings, int $type): bool
-    {
-        if ($subStrings === []) {
-            return true;
-        }
-
-        return array_any($subStrings, static fn (string $string) => ($type === self::PREFIX && str_starts_with($filename, $string)) ||
-            ($type === self::SUFFIX && str_ends_with($filename, $string)));
     }
 }